LiveUsb1.sh, change assignment to $count_gh_auth_checkmarks
The command substitution, `gh auth status |& grep --count ...` was counting zero checkmarks.
It just so happens that when `grep --count` finds zero of anything, the return code is not equal to zero.
This was triggering errexit.
Altered `grep --count` to `grep --only ... | wc -l`
Verified on the CLI, that `wc` will exit with code 0 if 0 checkmarks are found.