LiveUsb1.sh, comment out most "END" comments
The functions all have xtrace comments at their beginnings and ends, which print only in xtrace, which mark the boundaries of the functions and which display the function nesting depth. I've put the END comments into the RETURN trap in a previous commit, so with this commit, I commented out the rest of the END comments, since the `$(( --fn_lvl ))` functionality was in effect duplicated.