LiveUsb1.sh, work in setup_gpg() on `find` commands
Add a `find` command searching for any files owned by root and `chown`ing them to ownership of $USER.
Prior to that, add a local variable, a `readarray`, a `find` command and a `[[` command testing the value of the new local variable. The variable is $problem_files ; the `find` command looks for any file (FSO`s) that are not owned by either $USER or root; if any such files are found, the script throws an error and exits.
Add an explanatory (xtrace) note to this effect.