LiveUsb1.sh, remove commented out `true` commands; save a codeblock
Each function has a $fn_bndry comment in the script which prints only in xtrace at the in-to and out-of boundaries of each function; the out-of boundary commands had each been added individually until I realized the RETURN trap could be used for the same thing. At that point I had commented out each of those out-of boundary commands, which were `true` commands. So with this commit I have just deleted them.
Save a section of experimental out-of boundary code by commenting it out and adding a note.