LiveUsb1.sh, far too many changes for one commit
I've got some quoting issues going on in this script. Essentially, the quotation characters are over-used.
Remove as many single quotes as possible.
For string and array assignments, if `declare -p` on the command line would enclose a string in double quotes, then follow suit in the script.
Within hashtag (script-only) comments, all single quotes that would be used as a part of the English language, replace with backslashes, which is a littl jarring to read, but it is okay.  Fewer contractions, but possessives are acceptible.
For array assignments,
...structure them more readably and do not use bracket expansions.
For xtrace comments, use double quotes where possible. They print as single quotes in xtrace anyway.
For command names in comments, use grave accents
For bash syntax that does not include command names, again, in comments