LiveUsb1.sh, change absolute command paths; remove unused...
Remove unused comment and code
Change all in-code absolute command paths, ie. "/bin/rm", etc, to "$( type -P rm )", in order to force a PATH search each time the command is called.
In the case of `kill`, the issue is that if you want to use the binary and not the builtin, then you have to specify the absolute path somehow, but do so without hard-coding it.
In the case of `rm`, it`s just a "go the extra mile on being overly cautious" kind of thing.
However, from what I recall reading in Linux Foundation`s LFS207, some Linux distrobutions may still keep /bin/ and /usr/bin/ as separate directories, so hard-coding an absolute path could cause an error, in theory anyway.