LiveUsb1.sh, add an alias and edit set()
The alias at the beginning of each function, there for readability, is _als_function_boundary_in_. It includes a `local -`; with the set() function, any execution of `builtin set` intended to effect a global scope must occur prior to any `local -` command.
Add an alias, _als_function_set_boundary_in_, that does not include a `local -` so that set(), when it executes `builtin set`, can affect a global scope.